Fix crash when showing menu "Credits" page
 with firmware 5.07.01.

The issue was caused by the enhancement to RESTORE which allows it
to take a variable instead of a literal string.

Fixed by removing duplication between the denizens variable and label.

To maintain compatibility with firmware 5.07.00 continue to use EXECUTE
to perform the RESTORE of the menu data instead of the new functionality.